refactor(word-spell): extract WordSpellLibrarySource sub-component
Carved the Level + Phonemes UI out of WordSpellSimpleConfigForm into a
reusable component so the Advanced modal can reuse the exact same UI for
parity (added in the next commit). Behaviour is unchanged.
